ddracechat.h is used by the client, so move it to src/game
where ddracecommands.h already is.
